requirements.txt
Flask==2.3.2
flask-sqlalchemy==2.0.16
cryptography==41.0.1
pandas==2.0.2
plotly==5.15.0